backend/video_generation/requirements.txt
# Video Generation Dependencies
moviepy==1.0.3
pillow==10.1.0
numpy==1.24.3
imageio-ffmpeg==0.4.9
imageio==2.34.0

# AI Model Dependencies
requests==2.31.0
python-dotenv==1.2.2

# Optional: For local Diffusers models (requires significant GPU memory)
diffusers==0.21.4
torch==2.1.0
transformers==4.30.2
accelerate==0.20.0
